requirements.txt
# Flask
Flask
Flask-SQLAlchemy

# Database
SQLAlchemy
psycopg2-binary

# Others
python-dotenv
